promjene

+++ Projekat/Posao.ba-BE/src/main/java/ba/posao/controllers/KorisnikController.java
@@ -137,9 +137,13 @@ public class KorisnikController {
 		}
 		
 		String passMD5=korisnikService.toMD5(pass.getPassword());
+		if (passMD5.equals(_korisnik.getPassword()))
+		{
+			korisnikService.removeKorisnici(id);
+			return ResponseEntity.status(HttpStatus.OK).body(true);
+		}
+		else return ResponseEntity.status(HttpStatus.NOT_ACCEPTABLE).body("Pogresan password");
 		
-		korisnikService.removeKorisnici(id);
-		return ResponseEntity.status(HttpStatus.OK).body(true);
     }
     
     @RequestMapping(path="/tip", method = RequestMethod.GET)
